+++ src/components/account/accountShop/accountMyShopping.js
@@ -4,12 +4,13 @@ import PeriodDropdown from "components/common/dropdown/periodDropdown";
 import MyReviewList from "../accountReview/myReviewList"; // request My review 말고 my product 같은걸로 바꿔서 props로 배송상태 주고 상태에 따라 띄우는거 어떤가
 
 import classes from "./accountMyShopping.module.css";
+import CommonPageTitle from "components/common/commPageTitle";
 
 function AccountMyShopping(props) {
   const [inqPeriod, setInqPeriod] = useState("");
   const [inqDeliState, setInqDeliState] = useState("");
-  const [showInqPeriod, setShowInqPeriod] = useState(false);
-  const [showInqDeliState, setshowInqDeliState] = useState(false);
+  // const [showInqPeriod, setShowInqPeriod] = useState(false);
+  // const [showInqDeliState, setshowInqDeliState] = useState(false);
 
   const DUMMY_DATA_UNREVIEWED = [
     {
@@ -33,13 +34,17 @@ function AccountMyShopping(props) {
   // 기간이 선택되면 바로 밑에 보여주기
   const createPeriodHandler = (data) => {
     setInqPeriod(data);
-    setShowInqPeriod(true);
+    console.log(inqPeriod); // for debug
+    // setShowInqPeriod(true);
   };
   // 기간이 선택되면 바로 밑에 보여주기
   const createDeliStateHandler = (data) => {
     setInqDeliState(data);
-    setshowInqDeliState(true);
+    console.log(inqDeliState); // for debug
+    // setshowInqDeliState(true);
   };
+
+  /* 선택값 밑에 띄우기 - 필요 없어서 없애기로 했는데, 나중에 비슷한 기능 쓸지도 모르니까 일단 남겨놓음
   // x 버튼을 누르면 선택한 기간값 사라짐
   const clearPeriodValue = () => {
     setShowInqPeriod(false);
@@ -48,11 +53,11 @@ function AccountMyShopping(props) {
   const clearDeliStateValue = () => {
     setshowInqDeliState(false);
   };
-
+  */
   return (
     <div className={classes.accountMyShopping_Layout}>
       <div className={classes.acctMyShopping_Inner}>
-        <p className={classes.acctMyShopping_Title}>나의쇼핑</p>
+        <CommonPageTitle title="나의쇼핑" />
         <hr className={classes.acctMyShopping_LineD} />
         <div className={classes.acctMyShopping_Main}>
           <p className={classes.acctMyShopping_SemiTitle}>주문배송내역 조회</p>
@@ -61,13 +66,13 @@ function AccountMyShopping(props) {
             <PeriodDropdown onCreate={createPeriodHandler} />
             <DeliStateDropdown onCreate={createDeliStateHandler} />
           </div>
+          {/* 선택값 밑에 띄우기 - 필요 없어서 없애기로 했는데, 나중에 비슷한 기능 쓸지도 모르니까 일단 남겨놓음
           <div className={classes.acctMyShopping_prdndelis}>
             {showInqPeriod && (
               <p className={classes.acctMyShopping_SelectedRslt}>
-                {/* 에러나서 일단 없앰, 나중에 정식으로 손보자 */}
-                {/* <p className={classes.acctMyShopping_selectedTextArea}> */}
-                {inqPeriod}
-                {/* </p> */}
+                <div className={classes.acctMyShopping_selectedTextArea}>
+                  {inqPeriod}
+                </div>
                 <button
                   className={classes.acctMyShopping_selectedRmvBtn}
                   onClick={clearPeriodValue}
@@ -78,9 +83,9 @@ function AccountMyShopping(props) {
             )}
             {showInqDeliState && (
               <p className={classes.acctMyShopping_SelectedRslt}>
-                {/* <p className={classes.acctMyShopping_selectedTextArea}> */}
-                {inqDeliState}
-                {/* </p> */}
+                <div className={classes.acctMyShopping_selectedTextArea}>
+                  {inqDeliState}
+                </div>
                 <button
                   className={classes.acctMyShopping_selectedRmvBtn}
                   onClick={clearDeliStateValue}
@@ -90,18 +95,17 @@ function AccountMyShopping(props) {
               </p>
             )}
           </div>
+            */}
           <p className={classes.acctMyShopping_SemiTitle}>주문상품</p>
           <div className={classes.acctMyShopping_orderProductDetail}>
             <p>000000000 | 2022.01.01.</p>
-            {/* qestion 00000000 이거 뭐지??? 상품 선택하면 그 상품의 주문번호, 날짜 인가?*/}
-            <p className={classes.accountMyShopping_showDetail}>
+            {/* <p className={classes.accountMyShopping_showDetail}>
               상세보기 {">"}
-            </p>
+            </p> */}
             {/** qestion 상세보기는 페이지?? 모달?? */}
           </div>
           <MyReviewList reviews={DUMMY_DATA_UNREVIEWED} />
-          <p className={classes.acctMyShopping_SemiTitle}>배송상태</p>
-          {/* request 컴포넌트화 해주세요 */}
+          {/* <p className={classes.acctMyShopping_SemiTitle}>배송상태</p>
           <div className={classes.deliTableArea}>
             <table className={classes.deliTable}>
               <thead>
@@ -119,7 +123,7 @@ function AccountMyShopping(props) {
                 </tr>
               </tbody>
             </table>
-          </div>
+          </div> */}
         </div>
       </div>
     </div>