+++ src/components/common/commPageSemiTitle.js
@@ -1,0 +1,17 @@
+import React from "react";
+import classes from "./commPageSemiTitle.module.css";
+
+/**
+ * 페이지 메인 TITLE component
+ * @param {String} semiTitle Main Title Name
+ * @returns
+ */
+function CommPageSemiTitle({ semiTitle }) {
+  return (
+    <div className={classes.semiTitle}>
+      <p>{semiTitle}</p>
+    </div>
+  );
+}
+
+export default CommPageSemiTitle;