+++ HISTORY.md
@@ -14,6 +14,26 @@
 2. components/main 에 존재했던 footer, navigation 폴더 분리
    - components/main/footer 와 components/main/navigation 으로 분리
 
+### 2022.01.13. 권예림
+
+1. 저번주 페이지 수정사항 반영
+   - 모달 테스트(http://localhost:3000/modaltest)에 'product sales link' 버튼 만들어서 올리기
+   - 기간, 배송상태 dropdown 밑에 띄워주는 기능 없애기
+   - 상세보기, 배송상태 부분 없애기
+2. 공통 컴포넌트 'Common Page Title' 생성 및 페이지 적용
+
+   - 각 페이지 Main Title 공통 컴포넌트
+   - 기존 코드는 주석처리 해뒀음
+   - App.js에 있는 페이지만 확인하고 바꾼거라 그 외 페이지는 필요하면 알아서!
+
+   <적용 페이지> (+각 페이지 css파일)
+
+   - signup.js
+   - accountMyShopping.js
+   - accountMyPage.js
+   - signupAdditional.js
+   - MyPresentPage.js
+
 ### 2022.01.09 김효진
 
 1. fetch 공통 변경 (공통 에러창 개발 해야함)